+++ stats/src/univariate/mixed.rs
@@ -1,10 +1,7 @@
 //! Mixed bootstrap
 
-use std::cmp;
-
 use float::Float;
-use num_cpus;
-use thread_scoped as thread;
+use rayon::prelude::*;
 
 use tuple::{Tuple, TupledDistributionsBuilder};
 use univariate::resamples::Resamples;
@@ -20,65 +17,42 @@ pub fn bootstrap<A, T, S>(
 where
     A: Float,
     S: Fn(&Sample<A>, &Sample<A>) -> T + Sync,
-    T: Tuple,
+    T: Tuple + Send,
     T::Distributions: Send,
     T::Builder: Send,
 {
-    let ncpus = num_cpus::get();
     let n_a = a.len();
     let n_b = b.len();
     let mut c = Vec::with_capacity(n_a + n_b);
     c.extend_from_slice(a);
     c.extend_from_slice(b);
+    let c = Sample::new(&c);
 
-    unsafe {
-        let c = Sample::new(&c);
-
-        // TODO need some sensible threshold to trigger the multi-threaded path
-        if ncpus > 1 && nresamples > n_a {
-            let granularity = nresamples / ncpus + 1;
-            let statistic = &statistic;
-
-            let chunks = (0..ncpus)
-                .map(|i| {
-                    let mut sub_distributions: T::Builder =
-                        TupledDistributionsBuilder::new(granularity);
-                    let offset = i * granularity;
-
-                    thread::scoped(move || {
-                        let end = cmp::min(offset + granularity, nresamples);
-                        let mut resamples = Resamples::new(c);
-
-                        for _ in offset..end {
-                            let resample = resamples.next();
-                            let a: &Sample<A> = Sample::new(&resample[..n_a]);
-                            let b: &Sample<A> = Sample::new(&resample[n_a..]);
-
-                            sub_distributions.push(statistic(a, b))
-                        }
-                        sub_distributions
-                    })
-                })
-                .collect::<Vec<_>>();
-
-            let mut builder: T::Builder = TupledDistributionsBuilder::new(nresamples);
-            for chunk in chunks {
-                builder.extend(&mut (chunk.join()));
-            }
-            builder.complete()
-        } else {
-            let mut resamples = Resamples::new(c);
-            let mut distributions: T::Builder = TupledDistributionsBuilder::new(nresamples);
-
-            for _ in 0..nresamples {
+    (0..nresamples)
+        .into_par_iter()
+        .map_init(
+            || Resamples::new(c),
+            |resamples, _| {
                 let resample = resamples.next();
                 let a: &Sample<A> = Sample::new(&resample[..n_a]);
                 let b: &Sample<A> = Sample::new(&resample[n_a..]);
 
-                distributions.push(statistic(a, b))
-            }
-
-            distributions.complete()
-        }
-    }
+                statistic(a, b)
+            },
+        )
+        .fold(
+            || T::Builder::new(0),
+            |mut sub_distributions, sample| {
+                sub_distributions.push(sample);
+                sub_distributions
+            },
+        )
+        .reduce(
+            || T::Builder::new(0),
+            |mut a, mut b| {
+                a.extend(&mut b);
+                a
+            },
+        )
+        .complete()
 }